Kernel structure and partial implementation.
Kernel uses a serial GCD queue to martial access to the physics
simulation that will be running on a separate thread.  This is to avoid
mutation of data structures mid simulation. Main methods in the kernel
that drive the physics simulation still need to be implemented and
debugged.  Need to start on the system component to finalise the
structure of the kernel.