Starting point for system interface
Basic kernel test completed with proof of concept.  Kernel code not
exactly optimal. I have decided to go with something as close to the
original arbor.js as possible for the "system" interface.  The
component code is flexible enough for different types of "system"
implementations, at some point I can create a system class tailored for
positioning UIViews or CALayers.