Demo templates and API update
Added template projects for the demos.  At this point, these projects
are largely empty with a few supporting resources added.  They will
serve as examples and unit tests for this code base.  Updated the API
for clarity.  The API still needs further review and debug properties
should be removed or placed into the debug rendering protocol.