Atlas demo implementation, system updates
This commit contains updates to the system while implementing the Atlas
demo.  Atlas demo looks the part, just needs to load the graph from
JSON and render the nodes during the simulation.  System parameters are
now implemented (updates are made to a copy, then sent back to the
system to update the simulation).  Energy now supports copying to
support the implementation.  Added high level header, PSArborTouch.h,
to the code base for ease of use.  Updated Kernel API for clarity.